+++ src/lib/messages.ts
@@ -5,7 +5,13 @@ export const messages = {
         upload: {
             success: 'Upload successful!',
             progress: 'Upload in progress...',
-        }
+        },
+        reset: {
+            password:{
+                success: 'Password reset link sent successfully!',
+                progress: 'Sending password reset link...',
+            }
+        },
     },
     error: {
         required: (field: string) =>  `${ field } is required.`,
@@ -31,6 +37,19 @@ export const messages = {
         token: {
             default: 'Token is invalid or expired.',
             null: 'Token is required.',
+        },
+        email: {
+            default: 'Email is invalid.',
+            null: 'Email is required.',
+            invalid: 'Invalid email address.',
+        },
+        reset: {
+            password: {
+                default: 'Password reset failed. Please try again.',
+                null: 'Password is required.',
+                minLength: (min: number) => `Password must be at least ${ min } characters long.`,
+                maxLength: (max: number) => `Password must be at most ${ max } characters long.`,
+            },
         }
     },
 };
